Several changes: Broke the classes into individual files for clarity. Added the new http client for Splunk HEC messages. Removed original splunk_handler and syslog_handler. Split the process of downloading files and pushing files to end point into two separate threads. Files are now downloaded and a dir/file watch was created. Each line will be read, sent and remove from the file and one the full file is sent, it will be deleted.
